Rk3399: it boots on the NanoPi-M4
External SD works! Of course, nothing can boot because
of missing DTB and ACPI, but - wow.

After screwing around with countless broken ways of creating
a booting SD or eMMC, I settled on the eflasher image from
http://wiki.friendlyarm.com/wiki/index.php/NanoPi_M4

If SD is in - it boots the eflasher. If SD isn't it, it
boots UEFI.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>